teacher edge case

+++ Model/TeacherLoader.php
@@ -51,13 +51,23 @@ class TeacherLoader
     public function deleteTeacher($id)
     {
         $con = Database::openConnection();
-        $handle2 = $con->prepare('UPDATE class SET teacher_id = NULL WHERE teacher_id = :id2');
-        $handle2->bindValue(':id2', $id);
-        $handle2->execute();
-        $handle = $con->prepare('DELETE FROM teacher WHERE teacher_id = :id');
-        $handle->bindValue(':id', $id);
-        $handle->execute();
+        $handle3 = $con->prepare('SELECT * FROM class WHERE teacher_id = :id3');
+        $handle3->bindValue(':id3', $id);
+        $handle3->execute();
+        $checkClass = $handle3->fetchAll();
+
+        if (!empty($checkClass)) {
+            return "Teacher is still assigned to a class";
+        } else {
+            $handle2 = $con->prepare('UPDATE class SET teacher_id = NULL WHERE teacher_id = :id2');
+            $handle2->bindValue(':id2', $id);
+            $handle2->execute();
+            $handle = $con->prepare('DELETE FROM teacher WHERE teacher_id = :id');
+            $handle->bindValue(':id', $id);
+            $handle->execute();
+        }
     }
+
 
     public function addTeacher($name, $email)
     {
